bf_App/requirements.txt
Pillow
numpy
pyfits
